buildfix.

+++ src/server/game/DungeonFinding/LFGQueue.cpp
@@ -439,7 +439,7 @@ namespace lfg
                 {
                     if (Player* player = ObjectAccessor::FindConnectedPlayer(itRoles->first))
                     {
-                        if (player->GetMapId() == dungeon->map)
+                        if (player->GetMapId() == static_cast<uint32>(dungeon->map))
                         {
                             if (InstanceScript* instance = player->GetInstanceScript())
                             {
